requirements.txt
Django==5.2.8

django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0


# Django API Documentation
django-rest-swagger==2.2.0
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.11

langchain==1.0.8
langchain-classic==1.0.0
pgvector==0.3.6
langchain-postgres==0.0.16
langchain-openai==1.0.3
langgraph==1.0.3
langchain-anthropic==1.0.4
langchain-google-genai==3.0.3


